WIP complex country bounds
