+++ package.json
@@ -13,11 +13,13 @@
         "node-schedule": "^1.3.2",
         "retry-axios": "^2.2.1",
         "ts-node": "^10.0.0",
-        "xml-stream": "^0.4.5"
+        "xml-stream": "^0.4.5",
+        "yargs": "^17.0.1"
     },
     "devDependencies": {
         "@types/node": "^15.12.4",
         "@types/node-schedule": "^1.3.1",
+        "@types/yargs": "^17.0.2",
         "eslint": "^7.29.0",
         "eslint-plugin-prettier": "^3.4.0",
         "prettier": "^2.3.2",
